+++ src/main/java/org/spongepowered/api/text/Text.java
@@ -36,7 +36,7 @@ import org.spongepowered.api.data.DataSerializable;
 import org.spongepowered.api.data.MemoryDataContainer;
 import org.spongepowered.api.data.Queries;
 import org.spongepowered.api.Sponge;
-import org.spongepowered.api.locale.Dictionary;
+import org.spongepowered.api.text.dictionary.Dictionary;
 import org.spongepowered.api.scoreboard.Score;
 import org.spongepowered.api.text.action.ClickAction;
 import org.spongepowered.api.text.action.HoverAction;
