+++ src/main/java/org/spongepowered/api/text/dictionary/DictionarySourceResolver.java
@@ -22,7 +22,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.spongepowered.api.locale;
+package org.spongepowered.api.text.dictionary;
 
 import static com.google.common.base.Preconditions.checkNotNull;
