+++ src/main/java/org/spongepowered/api/text/dictionary/bundle/package-info.java
@@ -23,4 +23,4 @@
  * THE SOFTWARE.
  */
 @org.spongepowered.api.util.annotation.NonnullByDefault
-package org.spongepowered.api.locale.bundle;
+package org.spongepowered.api.text.dictionary.bundle;
