+++ src/main/java/org/spongepowered/api/text/dictionary/config/ConfigDictionary.java
@@ -22,7 +22,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.spongepowered.api.locale.config;
+package org.spongepowered.api.text.dictionary.config;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
@@ -30,9 +30,9 @@ import com.google.common.base.Objects;
 import ninja.leaping.configurate.ConfigurationNode;
 import ninja.leaping.configurate.hocon.HoconConfigurationLoader;
 import ninja.leaping.configurate.loader.ConfigurationLoader;
-import org.spongepowered.api.locale.CachingDictionary;
-import org.spongepowered.api.locale.RemoteDictionary;
-import org.spongepowered.api.locale.bundle.ResourceBundleDictionary;
+import org.spongepowered.api.text.dictionary.CachingDictionary;
+import org.spongepowered.api.text.dictionary.RemoteDictionary;
+import org.spongepowered.api.text.dictionary.bundle.ResourceBundleDictionary;
 
 import java.io.BufferedReader;
 import java.io.IOException;
