+++ src/main/java/org/spongepowered/api/text/dictionary/package-info.java
@@ -23,4 +23,4 @@
  * THE SOFTWARE.
  */
 @org.spongepowered.api.util.annotation.NonnullByDefault
-package org.spongepowered.api.locale.config;
+package org.spongepowered.api.text.dictionary;
